chore: set validation and typehint
